New smart antispam plugin
When bot is spammed, restirect that user to use this bot.
When user is still spamming, notice owner to kick ass that user.
When bot is spammed in group, bot will banned that user. But when bot is not admin, bot will leave that chat.

This plugin is private for prevent retarted user exploit it.

Signed-off-by: Ayra Hikari <AyraHikari@gmail.com>