Welcome: Now have two method script for security button
One is when muted user is click security button in any welcome msg,
when that user is previous muted by security welcome, unmute that user.

And one is when user is click security button in their welcome msg,
bot will edit and remove that button in welc msg, and unmute that user.

Signed-off-by: Ayra Hikari <AyraHikari@gmail.com>